modify catalogue script
I changed the afterbackup script for the catalogue backup role.
This script manages the backup of the *.bsr files to github.com. I
modified the script to only worry about the two bsr file,
BackupCatalog.bsr and bacula-fd.bsr.

This will get aroung the issues We were having when other *.bsr files
other than BackupCatalog.bsr and bacula-fd.bsr were moved/renamed or
deleted. The script was unable to handle the case of a file being
removed from the git repository.

Signed-off-by: Joseph A. Lutz <joseph.lutz@novatechweb.com>