Remove inventory variables from credential file paths.
Unnecessary, and was causing issues with recursively created
directories of credentials. Including the inventory hostname in
the credential path also complicates using a credential for a service
located on another host, as both inventory hosts would have a different
version of the credential.

Using playbook_dir as the credential location solves these issues.