Add new role to generate TLS certificates for Buildbot
Extracts certificate generation from master and worker roles. It's really
messy to generate a certificate authority, certificates for multiple Ansible
hosts and to do it securely.

Creating a separate role for this also opens the opportunity for certificates
to come from another source like an online CA or a credential management
system like Ansible Vault or Hashicorp Vault.