Change haproxy ACLs to use domain matching
Previously we used exact string matching on the Host header. This has a problem
if a client issues a request like http://domain:port because the port is 
included in the Host header. We can avoid this error by using a pattern match 
against just the dot-delimited portion of Host.